media: sun6i-csi: implement vidioc_enum_framesizes
Create sun6i_csi_capture_enum_framesizes which defines the min
and max frame sizes.

As per the specification of vidioc_enum_framesizes, if the supplied
index is >0 or no valid formats are found for the supplied format,
return -EINVAL.

Set the type to V4L2_FRMSIZE_TYPE_CONTINUOUS and set the min and max
width and height from the approriate macros, 32x32 - 4800x4800.

Signed-off-by: Adam Pigg <adam@piggz.co.uk>